Add filters to the top of the engagements archive.
Add basic search and taxonomy filter checkboxes to the top of the
engagements archive. Applying filters returns you to the engagements
archive filtered down to matching engagements.
• User can choose filters, which are submitted in standard GET format
back to the archive page.
• At `template_redirect`, we clean up the GET string so that arrays are
converted to comma-separated strings (solely for prettiness of URL).
• At the archive’s `pre_get_posts` action, we apply any user-selected
filters to the `WP_Query` so that the results match the user’s choices.
• Update the JS and CSS to handle the show/hide toggle