Populate the "about" content field.
We’re inserting the post_content form field in using CMB, which is kind
of a weird use. So, CMB, which typically deals with post_meta, isn’t
finding a value to insert in the content field. We have to add a filter
to populate the content, since it’s coming from the post_content, and
not from a meta value.