Add custom pre-processing to REST API, and geography selection prompts
Filter the incoming 'muext_geoid' parameter. Essentially, if the user chooses "Columbia," we want to include all engagements that apply to Columbia, like ZIP 65201 and Boone County and the schools district. Individual engagements are tagged with their service area, not every possible geoid that could apply.

Add geography selection prompt - a popup on the map, and updated list prompt